+++ govtool/frontend/src/context/wallet.tsx
@@ -209,8 +209,6 @@ const CardanoProvider = (props: Props) => {
       // return changeAddress for the usage of the pillars;
       return changeAddress;
     } catch (err) {
-      Sentry.setTag("wallet-action", "getChangeAddress");
-      Sentry.captureException(err);
       console.error(err);
     }
   };
@@ -225,8 +223,6 @@ const CardanoProvider = (props: Props) => {
       ).to_bech32();
       setWalletState((prev) => ({ ...prev, usedAddress }));
     } catch (err) {
-      Sentry.setTag("wallet-action", "getUsedAddresses");
-      Sentry.captureException(err);
       console.error(err);
     }
   };
@@ -264,12 +260,6 @@ const CardanoProvider = (props: Props) => {
               return enabledWalletApi;
             })
             .catch((e) => {
-              Sentry.addBreadcrumb({
-                category: "wallet",
-                message: "Wallet connection failed",
-                level: "warning",
-              });
-              Sentry.captureException(e, { data: window.cardano[walletName] });
               throw e.info;
             });
           await getChangeAddress(enabledApi);
@@ -366,8 +356,6 @@ const CardanoProvider = (props: Props) => {
 
           return { status: t("ok"), stakeKey: stakeKeySet };
         } catch (e) {
-          Sentry.setTag("wallet-action", "enable");
-          Sentry.captureException(e);
           console.error(e);
           setError(`${e}`);
           setAddress(undefined);
@@ -589,8 +577,6 @@ const CardanoProvider = (props: Props) => {
           disconnectWallet();
         }
 
-        Sentry.setTag("wallet-action", "buildSignSubmitConwayCertTx");
-        Sentry.captureException(error);
         console.error(error, "error");
         throw error?.info ?? error;
       }
@@ -639,8 +625,6 @@ const CardanoProvider = (props: Props) => {
 
         return certBuilder;
       } catch (e) {
-        Sentry.setTag("wallet-action", "buildVoteDelegationCert");
-        Sentry.captureException(e);
         console.error(e);
         throw e;
       }
@@ -678,8 +662,6 @@ const CardanoProvider = (props: Props) => {
         }
         return Certificate.new_drep_registration(dRepRegCert);
       } catch (e) {
-        Sentry.setTag("wallet-action", "buildDRepRegCert");
-        Sentry.captureException(e);
         console.error(e);
         throw e;
       }
@@ -708,8 +690,6 @@ const CardanoProvider = (props: Props) => {
         }
         return Certificate.new_drep_update(dRepUpdateCert);
       } catch (e) {
-        Sentry.setTag("wallet-action", "buildDRepUpdateCert");
-        Sentry.captureException(e);
         console.error(e);
         throw e;
       }
@@ -731,8 +711,6 @@ const CardanoProvider = (props: Props) => {
 
         return Certificate.new_drep_deregistration(dRepRetirementCert);
       } catch (e) {
-        Sentry.setTag("wallet-action", "buildDRepRetirementCert");
-        Sentry.captureException(e);
         console.error(e);
         throw e;
       }
@@ -785,8 +763,6 @@ const CardanoProvider = (props: Props) => {
 
         return votingBuilder;
       } catch (e) {
-        Sentry.setTag("wallet-action", "buildVote");
-        Sentry.captureException(e);
         console.error(e);
         throw e;
       }
@@ -998,8 +974,6 @@ function useCardano() {
         // TODO: type error
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } catch (e: any) {
-        Sentry.setTag("wallet-action", "enable");
-        Sentry.captureException(e);
         await context.disconnectWallet();
         navigate(PATHS.home);
         openModal({